refactor(relay): Slack/Feishu HTTP ingest + ingress routes move into their platform modules (S3)
Pure git mv + import-specifier updates; zero logic change. The class (a)
whole-module moves the audit prescribed, sequenced AFTER the plugin seam
(#569) took over construction and the interaction funnels so the move is a
move — route-seam adoption of plugin.verify/handle is the next PR (§16: never
move a body while also changing what it reaches).

slack-http-ingest.ts(+test) → platforms/slack/http-ingest.ts
slack-http-ingress.ts(+test) → platforms/slack/http-ingress.ts
feishu-http-ingest.ts → platforms/feishu/http-ingest.ts
feishu-http-ingress.ts(+test) → platforms/feishu/http-ingress.ts

Relay suite 417 passed unchanged.
