+++ packages/relay/src/index.ts
@@ -20,8 +20,8 @@ import { createRelayBrowserServer } from './relay-browser-server.js'
 import { WebchatRouter } from './webchat-router.js'
 import { RelayIngressManager } from './relay-ingress-manager.js'
 import { SlackEventDedup } from './slack-event-dedup.js'
-import { registerSlackHttpIngress } from './slack-http-ingress.js'
-import { registerFeishuHttpIngress } from './feishu-http-ingress.js'
+import { registerSlackHttpIngress } from './platforms/slack/http-ingress.js'
+import { registerFeishuHttpIngress } from './platforms/feishu/http-ingress.js'
 import { CollaborationRouter } from './collaboration-router.js'
 import { createAgentMsgRouter } from './agent-msg-router.js'
 import { mapAgentDirectory, toBotAssignment } from './bot-arbitration.js'
