+++ packages/relay/src/platforms/slack/http-ingest.test.ts
@@ -20,8 +20,8 @@ import {
   SlackHttpIngest,
   type SlackHttpIngestDeps,
   type SlackInteractiveBody
-} from './slack-http-ingest.js'
-import { verifySlackSignature } from './hooks/signature.js'
+} from './http-ingest.js'
+import { verifySlackSignature } from '../../hooks/signature.js'
 
 const AGENT_ID = '11111111-1111-4111-8111-111111111111'
 const INTEGRATION_ID = '22222222-2222-4222-8222-222222222222'
